license-server/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
exceptiongroup==1.3.0
fastapi==0.115.12
h11==0.16.0
idna==3.10
pydantic==2.11.7
pydantic_core==2.33.2
PyMySQL==1.1.1
python-dotenv==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
passlib==1.7.4  # <-- เพิ่มนี้สำหรับการเข้ารหัส password
bcrypt==4.1.2   # <-- เพิ่มนี้สำหรับการ hash แบบปลอดภัย
python-jose==3.3.0  # <-- เพิ่มนี้สำหรับ JWT (ถ้าใช้)
pymysql==1.1.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic[email]
requestsrequests==2.32.5
sse-starlette==2.1.0  
